+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/state/plugin/logic/channel/internal/ChannelStateLogic.kt
@@ -418,11 +418,13 @@ internal class ChannelStateLogic(
      *
      * @param memberUserId Updated member user id.
      * @param banned Shows whether a user is banned or not in this channel.
+     * @param banExpires The date when the ban expires.
      * @param shadow Shows whether a user is shadow banned or not in this channel.
      */
     fun updateMemberBanned(
         memberUserId: String?,
         banned: Boolean,
+        banExpires: Date?,
         shadow: Boolean,
     ) {
         mutableState.upsertMembers(
@@ -430,6 +432,7 @@ internal class ChannelStateLogic(
                 when (member.user.id == memberUserId) {
                     true -> member.copy(
                         banned = banned,
+                        banExpires = banExpires,
                         shadowBanned = shadow,
                     )
                     false -> member