+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/ui/common/feature/messages/list/MessageListController.kt
@@ -234,10 +234,11 @@ public class MessageListController(
         .flatMapLatest { state ->
             combine(
                 state.channelData,
+                state.members,
                 state.membersCount,
                 state.watcherCount,
                 state.pinnedMessages,
-            ) { _, _, _, _ ->
+            ) { _, _, _, _, _ ->
                 state.toChannel()
             }
         }