+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 
 ## stream-chat-android-ui-common
 ### 🐞 Fixed
+- Fixed `MessageListController.channel.members` not being updated when a member is banned. [#5368](https://github.com/GetStream/stream-chat-android/pull/5368)
 
 ### ⬆️ Improved
 
@@ -58,6 +59,7 @@
 ## stream-chat-android-ui-components
 ### 🐞 Fixed
 - Message list is cleared when the channel remove all its messages. [#5360](https://github.com/GetStream/stream-chat-android/pull/5360)
+- Ensure channels list doesn't show deleted messages on the last message preview. [#5367](https://github.com/GetStream/stream-chat-android/pull/5367)
 
 ### ⬆️ Improved
 
@@ -72,6 +74,9 @@
 ## stream-chat-android-compose
 ### 🐞 Fixed
 - Message list is cleared when the channel remove all its messages. [#5360](https://github.com/GetStream/stream-chat-android/pull/5360)
+- Increase max allowed votes base on available options. [#5363](https://github.com/GetStream/stream-chat-android/pull/5363)
+- Poll Attachment show description instead of question twice. [#5364](https://github.com/GetStream/stream-chat-android/pull/5364)
+- Show proper votes count in anonymous polls. [#5366](https://github.com/GetStream/stream-chat-android/pull/5366)
 
 ### ⬆️ Improved